refactor: move AccessorReactiveObserver to own file
