+++ packages/@lwc/engine-core/src/framework/decorators/api.ts
@@ -5,19 +5,14 @@
  * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/MIT
  */
 import features from '@lwc/features';
-import { assert, isFalse, isFunction, isTrue, isUndefined, toString } from '@lwc/shared';
+import { assert, isFunction, isUndefined, toString } from '@lwc/shared';
 import { logError } from '../../shared/logger';
 import { isInvokingRender, isBeingConstructed } from '../invoker';
-import {
-    componentValueObserved,
-    componentValueMutated,
-    ReactiveObserver,
-    JobFunction,
-} from '../mutation-tracker';
+import { componentValueObserved, componentValueMutated } from '../mutation-tracker';
 import { LightningElement } from '../base-lightning-element';
-import { getAssociatedVM, rerenderVM, VM } from '../vm';
-import { addCallbackToNextTick } from '../utils';
+import { getAssociatedVM } from '../vm';
 import { isUpdatingTemplate, getVMBeingRendered } from '../template';
+import { createAccessorReactiveObserver } from '../accessor-reactive-observer';
 
 /**
  * @api decorator to mark public fields and public methods in
@@ -74,57 +69,6 @@ export function createPublicPropertyDescriptor(key: string): PropertyDescriptor 
         enumerable: true,
         configurable: true,
     };
-}
-
-const DUMMY_ACCESSOR_REACTIVE_OBSERVER = {
-    observe(job: JobFunction) {
-        job();
-    },
-    reset() {},
-    link() {},
-} as unknown as AccessorReactiveObserver;
-
-function createAccessorReactiveObserver(vm: VM, set: (v: any) => void) {
-    // On the server side, we don't need mutation tracking. Skipping it improves performance.
-    return process.env.IS_BROWSER
-        ? new AccessorReactiveObserver(vm, set)
-        : DUMMY_ACCESSOR_REACTIVE_OBSERVER;
-}
-
-export class AccessorReactiveObserver extends ReactiveObserver {
-    private value: any;
-    private debouncing: boolean = false;
-    constructor(vm: VM, set: (v: any) => void) {
-        super(() => {
-            if (isFalse(this.debouncing)) {
-                this.debouncing = true;
-                addCallbackToNextTick(() => {
-                    if (isTrue(this.debouncing)) {
-                        const { value } = this;
-                        const { isDirty: dirtyStateBeforeSetterCall, component, idx } = vm;
-                        set.call(component, value);
-                        // de-bouncing after the call to the original setter to prevent
-                        // infinity loop if the setter itself is mutating things that
-                        // were accessed during the previous invocation.
-                        this.debouncing = false;
-                        if (isTrue(vm.isDirty) && isFalse(dirtyStateBeforeSetterCall) && idx > 0) {
-                            // immediate rehydration due to a setter driven mutation, otherwise
-                            // the component will get rendered on the second tick, which it is not
-                            // desirable.
-                            rerenderVM(vm);
-                        }
-                    }
-                });
-            }
-        });
-    }
-    reset(value?: any) {
-        super.reset();
-        this.debouncing = false;
-        if (arguments.length > 0) {
-            this.value = value;
-        }
-    }
 }
 
 export function createPublicAccessorDescriptor(
